refactor: extract config loading to break demo↔docguard import cycle
The Architecture validator flagged a real circular dependency:
demo.mjs → docguard.mjs (loadConfig) and docguard.mjs → demo.mjs (runDemo,
a statically-imported command). Since the validator counts dynamic imports
too, lazy-loading wouldn't help — the dependency had to actually go.

Moves loadConfig + its private helpers (autoDetectProjectType,
getProjectTypeDefaults, normalizeConfig, deepMerge, validator-key
normalizers) verbatim into a new cli/config.mjs that imports only shared.mjs
+ shared-ignore.mjs (never a command), so the graph is acyclic. docguard.mjs
and demo.mjs both import loadConfig from there. Byte-for-byte move — no logic
change.

Also:
- Fixed a stale "(23 validators)" count in `docguard --help` (now count-free;
  metrics-consistency only scans .md so it never caught this one).
- .gitignore *.vsix (vsce package output), alongside the existing *.tgz.

Validation: 632/632 tests pass; demo command works (loads fixture config via
config.mjs); guard's Architecture check goes from "circular dependency" to
N/A. No version bump — staged for the v0.23.0 release.
